Move SearchForm.js to SearchForm folder. Update behavior. Add testing
